Introduce common base entity for Comelit bridge
